fix some bugs about sql_type (#7803)
Approved by: @aressu1985, @heni02, @ouyuanning, @daviszhen, @nnsgmsone
