+++ pkg/sql/compile/compile.go
@@ -609,11 +609,11 @@ func (c *Compile) compileExternScan(ctx context.Context, n *plan.Node) ([]*Scope
 		return nil, err
 	}
 	if param.ScanType == tree.S3 {
-		if err := external.InitS3Param(param); err != nil {
+		if err := plan2.InitS3Param(param); err != nil {
 			return nil, err
 		}
 	} else {
-		if err := external.InitInfileParam(param); err != nil {
+		if err := plan2.InitInfileParam(param); err != nil {
 			return nil, err
 		}
 	}
@@ -633,7 +633,7 @@ func (c *Compile) compileExternScan(ctx context.Context, n *plan.Node) ([]*Scope
 				fileList[i] = strings.TrimSpace(fileList[i])
 			}
 		} else {
-			fileList, fileSize, err = external.ReadDir(param)
+			fileList, fileSize, err = plan2.ReadDir(param)
 			if err != nil {
 				return nil, err
 			}
