+++ pkg/sql/plan/build_load.go
@@ -15,6 +15,7 @@
 package plan
 
 import (
+	"context"
 	"encoding/json"
 	"strings"
 
@@ -26,6 +27,10 @@ import (
 )
 
 func buildLoad(stmt *tree.Load, ctx CompilerContext) (*Plan, error) {
+	if err := checkFileExist(stmt.Param); err != nil {
+		return nil, err
+	}
+
 	if err := InitNullMap(stmt.Param, ctx); err != nil {
 		return nil, err
 	}
@@ -125,6 +130,29 @@ func buildLoad(stmt *tree.Load, ctx CompilerContext) (*Plan, error) {
 	}
 	pn.GetQuery().LoadTag = true
 	return pn, nil
+}
+
+func checkFileExist(param *tree.ExternParam) error {
+	param.Ctx = context.TODO()
+	if param.ScanType == tree.S3 {
+		if err := InitS3Param(param); err != nil {
+			return err
+		}
+	} else {
+		if err := InitInfileParam(param); err != nil {
+			return err
+		}
+	}
+
+	fileList, _, err := ReadDir(param)
+	if err != nil {
+		return err
+	}
+	if len(fileList) == 0 {
+		return moerr.NewInvalidInput(param.Ctx, "the file does not exist in load flow")
+	}
+	param.Ctx = nil
+	return nil
 }
 
 func GetProjectNode(stmt *tree.Load, ctx CompilerContext, node *plan.Node, Name2ColIndex map[string]int32, clusterTable *ClusterTable) error {
